+++ src/pydidas/widgets/parameter_config/parameter_widgets_mixin.py
@@ -21,7 +21,7 @@ Parameters.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023 - 2025, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2026, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -61,7 +61,7 @@ class ParameterWidgetsMixIn:
 
         Parameters
         ----------
-        param : Parameter | str
+        param : Parameter or str
             A Parameter class instance.
         **kwargs : Any
             Optional keyword arguments. Supported keys are:
@@ -85,10 +85,10 @@ class ParameterWidgetsMixIn:
             halign_text : QtCore.Qt.Alignment, optional
                 The horizontal alignment for the text (label) widget. The default
                 is QtCore.Qt.AlignRight.
-            parent_widget : Union[QWidget, str, None], optional
-                The widget to which the label is added. If a string, this picks up the
-                calling class's ._widgets dictionary and selects the string key's value.
-                The default is self.
+            parent_widget : QWidget or str or None, optional
+                The widget to which the label is added. If a string,
+                this picks up the calling class's ._widgets dictionary
+                and selects the string key's value. The default is self.
         """
         _parent = kwargs.get("parent_widget", self)
         if isinstance(_parent, str):
@@ -198,8 +198,8 @@ class ParameterWidgetsMixIn:
         value : Any
             The new value for the Parameter.
         choices : None or Sequence[Any]
-            The new list of choices for the Parameter. If None, the choices
-            for the Parameter will be disabled.
+            The new list of choices for the Parameter. If None, the
+            choices for the Parameter will be disabled.
         emit_signal : bool
             Flag to toggle emitting a changed signal after updating the value
             and choices (if the value has changed). The default is True.